% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect.R
\name{bucher}
\alias{bucher}
\title{Simple indirect (Bucher) meta-analysis}
\usage{
bucher(abTE, se.abTE, cbTE, se.cbTE, effect, model, continuous, intervention,
  comparator, common, backtransf = FALSE, ab.studies, cb.studies)
}
\arguments{
\item{abTE}{Numeric. The treatment effect for a vs b. e.g. log OR, log HR,
mean difference}

\item{se.abTE}{Numeric. The standard error of the treatment effect for
a vs b, e.g. se of log OR}

\item{cbTE}{Numeric. The treatment effect for c vs b. e.g. log OR, log HR,
mean difference}

\item{se.cbTE}{Numeric. The standard error of the treatment effect for c vs
b, e.g. se of log OR}

\item{effect}{Character string describing the effect measure, e.g. 'Rate
Ratio' or 'log Odds Ratio'}

\item{model}{Character string indicating whether abTE and cbTE come from a
fixed effect model or a random effect model}

\item{continuous}{Logical (TRUE/FALSE) indicating whether the effect measure
is continuous (mean difference) or a ratio measure (odds ratio, hazard
ratio etc).}

\item{intervention}{Character string. Name of the intervention treatment}

\item{comparator}{Character string. Name of the comparator treatment}

\item{common}{Character string. Name of the common comparator that links the
indirect comparison, e.g. placebo}

\item{backtransf}{Logical indicating whether the results should be
exponentiated or not. If abTE and cbTE are on the log scale (e.g. log
hazard ratio) set this to TRUE to return the exponentiated results (e.g.
hazard ratio). If TRUE this will return both the log estimates and the
exponentiated estimates}

\item{ab.studies}{Character string giving the names of studies in the A vs B
comparison separated by commas}

\item{cb.studies}{Character string giving the names of studies in the C vs B
comparison separated by commas}
}
\description{
Simple indirect (Bucher) meta-analysis
}
\details{
Calculate an indirect estimate of the relative effect of two
  treatments using the Bucher method

  The inputs are the relative treatment effects for two pairs of treatments
  linked by a common comparator. For example, if you have the relative
  effects (e.g. log hazard ratio) for treatment A vs placebo and treatment C
  vs placebo then this function will return the relative effect of treatment
  A compared to treatment C.

  This function is mainly intended to be called from \code{\link{doBucher}} but can
  be used directly if required.

  \code{effect}, \code{model}, \code{intervention}, \code{comparator} and
  \code{common} are used for labels only. The results depend only on
  \code{abTE}, \code{cbTE} and the respective standard errors.

  @return A data frame with the following columns:
\itemize{
 \item \code{Intervention} The name of the intervention
 \item \code{Comparator} The name of the comparator
 \item \code{Common} The name of the common treatment linking intervention
 and comparator
 \item \code{Effect} The type of effect measure. Takes the
   value of the \code{effect} argument
 \item \code{Model} The type of model. Should be 'Fixed', 'Random' or NA
 \item \code{log.TE.ind} The treatment effect on log scale, e.g. log OR
 \item \code{log.lower.ind}, \code{log.upper.ind} The upper and lower 95\% confidence
   intervals for the log treatment effect
 \item \code{se.log.TE.ind} The standard error for the log treatment effect
 \item \code{TE.ind}, \code{lower.ind}, \code{upper.ind} The treatment effect with lower
   and upper confidence intervals backtransformed to a linear scale
 \item \code{n.studies} The number of studies included in the analysis
 \item \code{Studies} The names of the studies included in the analysis
}

  @seealso \code{\link{doBucher}}
}

