% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{calcAllPairs}
\alias{calcAllPairs}
\title{Calculate all pairwise treatment effects in a network meta-analysis}
\usage{
calcAllPairs(mtcRes, expon = FALSE, ...)
}
\arguments{
\item{mtcRes}{An object of class \code{mtc.results} as returned by mtc.run}

\item{expon}{Logical indicating whether the results should be exponentiated
or not. If your analysis is based on log hazard ratio or log odds ratio etc
then set this to TRUE to get the results as the corresponding hazard ratio
or odds ratio.}

\item{...}{Additional arguments passed down from extractMTCResults}
}
\value{
A data frame reporting summary statistics (mean, median etc) for the
  posterior distribution of each treatment comparison. Each row in the data
  frame represents one treatment comparison.
}
\description{
Calculate all pairwise treatment effects in a network meta-analysis
}
\details{
This function takes a \code{mtc.results} object and extracts summary
  statistics for all pairwise treatment comparisons in the network using
  \code{relative.effect} from gemtc. This function should mainly be used via
  \code{extractMTCresults} but can be called directly on any
  \code{mtc.results} object. Only results for the treatment comparisons are
  returned, additional nodes such as hetereogeneity parameters are not
  returned. Model comparisons statistics and heterogeneity parameters can be
  obtained using \code{extractModelFit}
}
\seealso{
\code{\link[gemtc]{mtc.run}}, \code{\link[gemtc]{relative.effect}}
  \code{\link{extractMTCResults}}, \code{\link{extractModelFit}}
}
