% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect.R
\name{doBucher}
\alias{doBucher}
\title{Perform all possible indirect comparisons for a given data set}
\usage{
doBucher(comparisons, direct, effectType = "all", effect_measure,
  backtransf = FALSE)
}
\arguments{
\item{comparisons}{A data frame with four columns: StudyName, study,
comparator, treatment. Describes the treatment comparisons present in the
dataset. study, comparator and treatment must be numbers. For example,
study = 4, comparator=1, treatment=2 represents the comparison of treatment
2 vs treatment 1 in study 4}

\item{direct}{A data frame containing the results of direct head-to-head
meta-analysis for the treatment comparisons of interest if only one study is
available for a given comparison then the result of that study should be
used. This data frame can be created by using \code{doDirectMeta} and
\code{extractDirectRes} in that order}

\item{effectType}{Character string indicating what type of results are
required. Default is 'all' which will return both fixed effect and random
effect results. Alternatives are 'Fixed' or 'Random' (Case sensitive) if
only one set of results is required}

\item{effect_measure}{A character string indicating what type of effect
measure is used, e.g. 'Rate Ratio', 'Odds Ratio' etc.}
}
\description{
Perform all possible indirect comparisons for a given data set
}
\details{
This function performs indirect meta-analysis for all possible
  comparisons in a given data set. This function takes a set of treatment
  comparisons from one or more studies and identifies all possible indirect
  comparisons where two treatments can be connected via a common comparator.
  If there is more than one way to connect two treatments then all possible
  variations are calculated. This function calls \code{\link{bucher}} internally
  to calculate the treatment effects

  The inputs for this function will usually be the results from direct
  meta-analysis for a given set of treatments. The recommended workflow is
  to use \code{\link{doDirectMeta}} to perform head to head meta-analysis for
  a given set of treatments, extract the results as a data frame using
  \code{\link{extractDirectRes}} then use that data frame to provide the inputs
  for this function.

  @return A data frame with the following columns:
\itemize{
 \item \code{Intervention} The name of the intervention
 \item \code{Comparator} The name of the comparator
 \item \code{Common} The name of the common treatment linking intervention
 and comparator
 \item \code{Effect} The type of effect measure. Takes the
   value of the \code{effect} argument
 \item \code{Model} The type of model. Fixed effect or Random Effects.
 \item \code{log.TE.ind} The treatment effect on log scale, e.g. log OR
 \item \code{log.lower.ind}, \code{log.upper.ind} The upper and lower 95\% confidence
   intervals for the log treatment effect
 \item \code{se.log.TE.ind} The standard error for the log treatment effect
 \item \code{TE.ind}, \code{lower.ind}, \code{upper.ind} The treatment effect with lower
   and upper confidence intervals backtransformed to a linear scale
 \item \code{n.studies} The number of studies included in the analysis
 \item \code{Studies} The names of the studies included in the analysis
}
}
\seealso{
\code{\link{bucher}}, \code{\link{doDirectMeta}},
  \code{\link{extractDirectRes}}
}

