% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct.R
\name{doDirectMeta}
\alias{doDirectMeta}
\title{Perform multiple direct head to head meta-analyses from a single data frame}
\usage{
doDirectMeta(df, effectCode, dataType, backtransf = FALSE, method = "MH")
}
\arguments{
\item{df}{A data frame as returned by \code{formatDataToDirectMA}}

\item{effectCode}{A character string indicating the underlying effect
estimate. This is used to set the \code{sm} argument of the underlying
analysis functions from the \code{meta} package. Acceptable values are
'RD', 'RR', 'OR', 'HR', 'ASD', 'MD', 'SMD'}

\item{dataType}{A character string specifying which type of data has been
provided. Currently only 'treatment difference' or 'binary are supported}

\item{backtransf}{A logical indicating whether results should be back
transformed. This is used to set the corresponding \code{backtransf}
argument of the underlying functions from the \code{meta} package. If
\code{backtransf=TRUE} then log odds ratios (or hazard ratios etc) will be
converted to odds ratios on plots and print outs}

\item{method}{A character string indicating which method is used for pooling
studies. This argument is only applicable if data_type='binary'. Must be one
of 'MH' (Default, Mantel-Haenszel method), 'inverse' or 'Peto'}
}
\value{
A data frame
}
\description{
Perform multiple direct head to head meta-analyses from a single data frame
}
\details{
This function provides a wrapper around the \code{metagen} or
  \code{metabin} functions from the \code{meta} package to one or more
  analyses to be carried out from a single data frame. This is most useful
  when direct meta-analysis is required to support all pairwise comparisons
  in a network meta-analyis or effect estimates are required for multiple
  pairs of treatments to perform simple indirect meta-analysis using the
  Bucher method
}
\seealso{
\code{\link{formatDataToDirectMA}}, \code{\link[meta]{metagen}},
  \code{\link[meta]{metabin}}
}

