% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct.R
\name{drawForest}
\alias{drawForest}
\title{Draw a forest plot}
\usage{
drawForest(meta, showFixed = TRUE, showRandom = TRUE, ...)
}
\arguments{
\item{meta}{an object of class c("metagen", "meta"), c("metacont", "meta") or
c("metabin", "meta") as returned by the functions \code{metagen},
\code{metabin} or \code{metacont} in the package meta}

\item{showFixed, showRandom}{Logical indicating whether fixed effect and/or
random effects results should be shown on the forest plot. By default both
are shown set the appropriate argument to FALSE if you want to exclude that
result from the plot.}

\item{...}{additional arguments to be passed to \code{forest}
e.g. \code{col.square='red'}, \code{col.diamond='black'}, \code{smlab='Odds
Ratio'}}
}
\description{
Draw a forest plot
}
\details{
This function provides a very simple wrapper around \code{forest}
  from the \code{meta} package. The main purpose is to allow some defaults
  to be set for the arguments to \code{forest} and to work out a reasonable
  scale for the x-axis automatically.

  By default pooled estimates for both fixed and random effects models will
  be shown. If there is only one study comparing a given pair of treatments
  then the results of that study are shown but no pooled estimates are
  displayed.

  @return NULL
}
\seealso{
\code{\link[meta]{forest}}
}
