% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{extractCoda}
\alias{extractCoda}
\title{Extract the coda from an mtc.result object}
\usage{
extractCoda(mtcResults, summarise = TRUE)
}
\arguments{
\item{mtcResults}{An \code{mtc.result} object as returned by \code{mtc.run}
from the gemtc package}

\item{summarise}{logical. If TRUE (default) and the mtc.result object
includes more than one chain then the mean across all chains is calculated
for each variable at each iteration of the chains}
}
\value{
If summarise = TRUE:
\itemize{
\item A data frame with one column per variable plus one column indicating
the iteration of the chain. The values for each variable will be the mean
across all chains for each iteration of the chain
}
If summarise = FALSE:
\itemize{
\item A data frame with one column per variable plus one column indicating
the iteration of each chain and one column indicating the chain numbered 1,
2, 3 etc. The values for each variable will be the value of the variable at
the indicated iteration of the indicated chain.
}
}
\description{
Extract the coda from an mtc.result object
}
\details{
This function takes the output from running an MTC model using
  mtc.run, extracts the coda from the \code{mtc.result} object and returns a
  data frame. If the summarise argument is TRUE and multiple chains were used
  for the MCMC then this function will return the mean across all chains for
  each variable in the model at each iteration of the chain. This is the
  default behaviour. If the summarise argument is FALSE then results for each
  individual chain are preserved.
}
\seealso{
\code{\link[gemtc]{mtc.run}}
}

