% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nma.R
\name{extractComparison}
\alias{extractComparison}
\title{Extract treatment comparison information from mtc summary output}
\usage{
extractComparison(df)
}
\arguments{
\item{df}{A data frame as returned by \code{calcAllPairs}}
}
\value{
A data frame with additional columns
}
\description{
Extract treatment comparison information from mtc summary output
}
\details{
In the summary output from a network meta-analysis treatment
  comparisons are commonly labelled as 'd.1.2' for the relative effect of
  treatment 2 compared to treatment 1. This function removes the 'd' and
  splits '1' and '2' into separate columns which is more useful for
  downstream reporting. This function is intended to work on the output from
  \code{calcAllPairs}. This functionality will usually be accessed via
  extractMTCResults and should only be used directly if you understand what
  you are doing.
}
\seealso{
\code{\link{calcAllPairs}}, \code{\link{extractMTCResults}}
}

