% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/direct.R
\name{extractDirectRes}
\alias{extractDirectRes}
\title{Extract summary results for a direct meta-analysis}
\usage{
extractDirectRes(metaRes, effect, intervention = "Int", comparator = "Con",
  interventionCode = NA, comparatorCode = NA, backtransf = FALSE)
}
\arguments{
\item{metaRes}{An object of class \code{c("metagen", "meta")} or c("metabin",
"meta"). These objects are lists containing the results of direct
meta-analysis. See \code{\link[meta]{metagen}}, \code{\link[meta]{metabin}}
for a description of exactly what is contained in the list}

\item{effect}{a character string describing the effect estimate, e.g. 'Rate
Ratio', 'Odds Ratio', 'Hazard Ratio'}

\item{intervention}{A character string describing the name of the
intervention. Defaults to 'Int' if not provided}

\item{comparator}{A character string describing the name of the comparator.
Defaults to 'Con' if not provided}

\item{backtransf}{A logical indicating whether results should be back
  transformed. If \code{backtransf=TRUE} then log odds ratios (or hazard
  ratios etc) will be converted to odds ratios on plots and print outs.}
}
\value{
A data frame with the following columns:
\itemize{
 \item \code{Intervention} The name of the intervention
 \item \code{InterventionCode} The ID number of the intervention in the
 current set of analyses. NA if not provided
 \item \code{Comparator} The name of the comparator
 \item \code{ComparatorCode} The ID number of the comparator in the current
   set of analyses
 \item \code{Effect} The type of effect measure. Takes the
   value of the \code{effect} argument
 \item \code{Model} The type of model. Fixed effect or Random Effects.
   Blank if there is only one study
 \item \code{log.TE} The treatment effect on log scale, e.g. log OR
 \item \code{seTE.log} The standard error for the log treatment effect
 \item \code{log.lower}, \code{log.upper} The upper and lower 95\% confidence
   intervals for the log treatment effect
 \item \code{z}, \code{p} The z-value and corresponding p-value for the test
   of effect
 \item \code{level} The level for the confidence intervals. Defaults to 0.95
   for a 95\% confidence interval
 \item \code{TE}, \code{lower}, \code{upper} The treatment effect with lower
   and upper confidence intervals backtransformed to a linear scale
 \item \code{Tau.sq} The heterogeneity variance
 \item \code{I.sq}, \code{I.sq.lower}, \code{I.sq.upper} The heterogeneity statistic
   I-squared with upper and lower confidence intervals.
}
}
\description{
Extract summary results for a direct meta-analysis
}
\details{
This function extracts the results of a meta-analysis from a
  list-type object produced by the \code{meta} package and returns them as a
  data frame. If there is only one study comparing two treatments then no
  meta-anlaysis is performed but the results of the primary study are
  included in the output. In this case the fields \code{Model},
  \code{Tau.sq}, \code{method.tau} and \code{I.sq} in the output will be
  blank as these fields have no meaning for a single study. This is useful if
  you want to use these results to perform simple indirect (Bucher) analyses.

  If more than one study is available then both fixed effect and random
  effects results will be returned
}
\seealso{
\code{\link[meta]{metagen}}, \code{\link[meta]{metabin}}
}

