% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{extractModelFit}
\alias{extractModelFit}
\title{Extract model comparison statistics from a \code{mtc.result} object}
\usage{
extractModelFit(mtcRes)
}
\arguments{
\item{mtcRes}{an object of class \code{mtc.result} as returned by
\code{mtc.run} from gemtc}
}
\value{
A data frame containing the model comparison statistics.
}
\description{
Extract model comparison statistics from a \code{mtc.result} object
}
\details{
This function returns model comparison statistics (DIC, pD, Dbar)
  from a \code{mtc.result} object. If the results object is derived from a
  random effects model then the between trials standard deviation will also
  be provided as sd.d. This function should mainly be used via
  \code{extractMTCResults} but can be used directly on a \code{mtc.result}
  object.
}
\seealso{
\code{\link{extractMTCResults}}, \code{\link[gemtc]{mtc.run}}
}
