% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nma.R
\name{extractNodesplit}
\alias{extractNodesplit}
\title{Extract node splitting results}
\usage{
extractNodesplit(ns.res, treatments, backtransf)
}
\arguments{
\item{ns.res}{An object of class \code{mtc.nodesplit}}

\item{treatments}{A data frame with columns 'description' defining the
treatment names and 'id' defining the treatment ID numbers.}

\item{back_calc}{A logical indicating whether results should be back transformed.
  If set to TRUE then log odds ratios (or hazard ratios etc) will
  be converted to odds ratios (or hazard ratios etc).}
}
\value{
A data frame. Column headers starting 'dir.' are the treatment effect
  and corresponding 95\% confidence intervals for the direct comparison. Column
  headers starting 'ind.' report the same information for the indirect
  comparison. Column headers starting 'cons.' report the pooled estimate
  using both direct and indirect evidence under a standard consistency model.
  Columns labelled RoR report the ratio of direct:indirect effect estimates
  and the corresponding 95\% confidence interval
}
\description{
Extract node splitting results
}
\details{
This function takes the output from the function mtc.nodesplit and
  returns a data frame of the results for each treatment comparison that can
  be split. The results are presented as one row per treatment comparison.
}
\seealso{
\code{\link[gemtc]{mtc.nodesplit}}
}

