% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extractSummary}
\alias{extractSummary}
\title{Extract essential summary estimates after fitting the Reitsma bivariate model}
\usage{
extractSummary(fit.reitsma)
}
\arguments{
\item{fit.reitsma}{An object of class \code{reitsma} returned by fitting the
Reitsma bivariate model using the function named \code{reitsma} in the mada
package}
}
\value{
A data frame containing the mean and 95% confidence intervals for the
  summary estimates of sensitivity, specificity, positive LR, negative LR and
  DOR.
}
\description{
Extract essential summary estimates after fitting the Reitsma bivariate model
}
\details{
This function takes the output from a fit of the Reitsma bivariate
  model, extracts the essential summary estimates and returns them in a
  convenient data frame. The Positive/Negative likelihood ratios and the
  diagnostic odds ratio are calculated using the method of Zwindermann and
  Bossuyt. This method uses an MCMC sampling based approach to calculate the
  results based on a fit of the bivariate model.
}
\seealso{
\code{\link[mada]{reitsma}}, \code{\link[mada]{SummaryPts}}
}

