% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nma.R
\name{extractTOI}
\alias{extractTOI}
\title{Extract all treatment comparisons involving a specific treatment if interest}
\usage{
extractTOI(df, treatments, toi, intervention = TRUE,
  reportOrder = "default")
}
\arguments{
\item{df}{A data frame as produced by \code{extractMTCResults}}

\item{treatments}{A data frame with two columns 'id' and 'description'. 'id'
must be the treatment id numbers corresponding to the way the treatments
were coded in the network. 'description' should be the name of the
treatment. An optional third column named 'Order' may also be provided. If
present this controls the order in which treatments are presented in the
output. The values should be a sequence of numbers indicating the order in
which the treatments should be sorted. If the 'Order' column is present
then set \code{orderResults=TRUE}}

\item{toi}{An integer specifying the id number of the treatment of interest
in the network}

\item{intervention}{Logical indicating whether the treatment of interest is
an intervention or a comparator (e.g. placebo). This controls which column
of \code{df} is searched to identify relevant treatment comparisons}

\item{reportOrder}{A character string indicating whether the treatments
  should be reported in the order the order they were provided (default) or
  if a custom order is required. Acceptable values are 'default' or
  'custom'. If this is set to 'custom' then \code{treatments} must contain a
  column named 'Order', see above.}
}
\value{
A data frame containing all treatment comparisons involving the
  specified treatment of interest.
}
\description{
Extract all treatment comparisons involving a specific treatment if interest
}
\details{
This is a simple filtering function to return all treatment
  comparisons involving a specific treatment of interest. This function is
  designed to be used with the output of \code{extractMTCResults}. There are
  two intended use cases:
  \enumerate{
  \item To return results for a given intervention of interest compared to
  all other treatments in the network
  \item To return results for all interventions compared to a given control
  treatment, e.g. placebo.
  }
}
\seealso{
\code{\link{extractMTCResults}}
}

