% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{forestmada2}
\alias{forestmada2}
\title{Extend the forestmada function with more flexible axes}
\usage{
forestmada2(x, ci, plotci = TRUE, main = "Forest plot", xlab = NULL,
  digits = 2L, snames = NULL, subset = NULL, pch = 15, cex = 1,
  cipoly = NULL, polycol = NA, xlim, ...)
}
\description{
Extend the forestmada function with more flexible axes
}
\details{
This function extends the forestmada function from the mada package
  to allow the user to control the range of the x-axes. The primary purpose
  is to allow the user to set \code{xlim = c(0,1)} in forest plots of
  sensitivity or specificity
}
\seealso{
\code{\link[mada]{forest}} for details of the other arguments
}
