% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{i2dacc}
\alias{i2dacc}
\title{Calculate I-squared for diagnostic accuracy parameters}
\usage{
i2dacc(df)
}
\arguments{
\item{df}{A data frame of 2x2 diagnostic accuracy data. The following
columns must be present 'StudyID, 'TP', 'TN', 'FP', 'FN' corresponding to
the number of True Positives, True Negatives, False Postives and False
Negatives respectively.}
}
\value{
A data frame of I-squared values with confidence intervals for each
  of the parameters.
}
\description{
Calculate I-squared for diagnostic accuracy parameters
}
\details{
This function will calculate the I-squared heterogeneity statistic
  with 95% confidence intervals for each of the following diagnostic accuracy
  parameters: Sensitivity, Specificity, Diagnostic Odds Ratio and both
  positive and negative likelihood ratios.

  The I-squared values are calculated using functions from the \code{meta}
  package but no pooling of the study estimates are performed. These
  functions are used only for convenience to provide the I-squared values.
  Sensitivity and specificity use \code{metaprop} for proportions. DOR,
  Positive LR and negative LR use \code{metagen} for generic inverse
  variance.
}
\seealso{
\code{\link[meta]{metaprop}}, \code{\link[meta]{metagen}}
}
