% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{makeTab}
\alias{makeTab}
\title{Generate a square grid reporting all pairwise comparisons of treatments in
the network}
\usage{
makeTab(results, coding, rounding = 2, reportOrder = "default", ...)
}
\arguments{
\item{results}{A data frame as returned by \code{nameTreatments}}

\item{coding}{A data frame with two columns 'id' and 'description'. 'id'
must be the treatment id numbers corresponding to the way the treatments
were coded in the network. 'description' should be the name of the
treatment. An optional third column named 'Order' may also be provided. If
present this controls the order in which treatments are presented in the
output. The values should be a sequence of numbers indicating the order in
which the treatments should be sorted. If the 'Order' column is present
then \code{reportOrder}  should be set to 'custom'}

\item{rounding}{The number of decimal places to be included in the output.
The default is two.}

\item{reportOrder}{A character string indicating whether the treatments
should be reported in the order the order they were provided (default) or
if a custom order is required. Acceptable values are 'default' or 'custom'.
If this is set to 'custom' then \code{coding} must contain a column named
'Order'. See above.}

\item{...}{Arguments passed from other functions, primarily
\code{extractMTCResults}}
}
\value{
A data frame
}
\description{
Generate a square grid reporting all pairwise comparisons of treatments in
the network
}
\details{
This function takes a data frame as returned by
  \code{nameTreatments} and returns a square grid containing all pairwise
  treatment comparisons in the network. The results are presented for column
  versus row. This is a common approach to reporting MTC results. This
  function should mainly be used via \code{extractMTCResults}
}
\seealso{
\code{\link{extractMTCResults}}, \code{\link{nameTreatments}}
}

