% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mautils.R
\docType{package}
\name{mautils}
\alias{mautils}
\alias{mautils-package}
\title{mautils: Utilities to facilitate an automated workflow for meta-analysis}
\description{
This package provides a collection of helper functions to support an
automated workflow for meta-analysis. This is mainly intended to be used in
situations where analyses are required for many different treatment
comparisons in a single project or multiple different meta-analysis models
need to be fitted for the same data set. The functions provided are mainly
simple wrappers around functions from the meta package for direct
meta-analysis or the gemtc package for network meta-analysis.
}

