% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nma.R
\name{nameTreatments}
\alias{nameTreatments}
\title{Match treatment names to ID numbers}
\usage{
nameTreatments(results, coding, ...)
}
\arguments{
\item{results}{A data frame as returned by \code{extractComparison}}

\item{coding}{A data frame with two columns 'id' and 'description'. 'id' must
  be the treatment id numbers corresponding to the way the treatments were
  coded in the network. 'description' should be the name of the treatment.}
}
\value{
The same data frame with the treatment names appended
}
\description{
Match treatment names to ID numbers
}
\details{
This function matches the coded treatment id numbers in the network
  to the corresponding human readable names. The mapping from id number to
  name should be provided as a two column data frame via the \code{coding}
  argument. This function is intended to work on the data frame generated as
  the output from \code{extractComparison}. This function will mainly be
  called via \code{extractMTCResults} and should only be used directly if you
  understand what you are doing. The general work flow is \code{mtc.run} >
  \code{calcAllPairs} > \code{extractComparison} > \code{nameTreatments} >
  \code{makeTab}. \code{extractMTCResults} will handle the last four steps
  for you.
}
\seealso{
\code{\link{extractComparison}}, \code{\link{calcAllPairs}}, \code{\link{extractMTCResults}}
}

