% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot.reitsma2}
\alias{plot.reitsma2}
\title{Extend the plot method for reitsma objects}
\usage{
\method{plot}{reitsma2}(x, extrapolate = FALSE, plotsumm = TRUE,
  level = 0.95, ylim = c(0, 1), xlim = c(0, 1), pch = 1, sroclty = 1,
  sroclwd = 1, predict = FALSE, predlty = 3, predlwd = 1,
  type = "ruttergatsonis", xlab, ...)
}
\description{
Extend the plot method for reitsma objects
}
\details{
This function extends the plot method for reitsma objects from the mada package
  to allow the user to set the label for the x-axis.
}
\seealso{
\code{\link[mada]{reitsma-class}} for details of the other arguments
}
