% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nma.R
\name{plotEstimates}
\alias{plotEstimates}
\title{Plot the results of a network meta-analysis}
\usage{
plotEstimates(df, yvar, xvar = "median", lowLimit = "CrI_lower",
  hiLimit = "CrI_upper", xlabel = "Effect Estimate", noEffectLine = 1,
  yOrder = NA)
}
\arguments{
\item{df}{A data frame containing summary statistics for the posterior
distribution of the parameters in a network meta-analysis.}

\item{yvar}{A character string specifying the name of the column in \code{df}
that contains the variable to be used on the y-axis. Usually this will be
the names of the treatments in the network.}

\item{xvar}{A character string specifying the name of the column in \code{df}
that contains the variable to be used on the x-axis. Usually this will be
the median of the posterior distribution of the treatment effect}

\item{lowLimit,}{hiLimit Character strings specifying the names of the
columns in \code{df} that specify the lower and upper limits of the 95%
credible intervals. If you have followed the work flow in this package
these will be called 'CrI_lower' and 'CrI_upper' respectively.}

\item{xlabel}{A character string to be used as a label on the x-axis. This is
usually the type of effect measure, e.g. 'Odds ratio', 'Rate Ratio' etc.}

\item{noEffectLine}{An integer indicating where to draw the line of no
effect. For ratio measures this is usually 1. For mean differences this is
usually 0.}

\item{yOrder}{A character string giving the name of the column in \code{df}
  that specifies the order of treatments on the y-axis. The default is NA in
  which case the treatments will be displayed on the y-axis alphabetically
  from top to bottom. If a specific order is required this is best specified
  in a column named 'Order' containing a series of numbers indicating the
  plotting order from top to bottom.}
}
\value{
A ggplot object
}
\description{
Plot the results of a network meta-analysis
}
\details{
This function should most commonly be used with the output from
  extractTOI to plot the results for a particular treatment of interest. This
  function uses ggplot to do the actual plotting and returns a ggplot object
  which may be further modified if required or just saved to a file.
}
\seealso{
\code{\link{extractTOI}}, \code{\link[ggplot2]{geom_point}},
  \code{\link[ggplot2]{geom_errorbarh}}
}

