% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{plotRanks}
\alias{plotRanks}
\title{Plot rank probabilities}
\usage{
plotRanks(ranks)
}
\arguments{
\item{ranks}{A data frame of ranking probabilities for each treatment as
returned by \code{extractRanks}}
}
\value{
A ggplot object which can then be saved using an appropriate graphics
  device, e.g. jpeg, png, pdf etc.
}
\description{
Plot rank probabilities
}
\details{
This function takes a data frame of ranking probabilities for each
  treatment in a network and constructs a 'rankogram'; i.e. a bar chart
  showing the probability of being ranked first, second, third etc. for each
  treatment
}
\seealso{
\code{\link{extractRanks}}, \code{\link[gemtc]{rank.probability}}
}
