% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plotSROC}
\alias{plotSROC}
\title{Draw a Summary ROC curve based on the results of the bivariate model}
\usage{
plotSROC(fit.reitsma, show.data = TRUE, show.prediction = TRUE,
  summary.legend.pos = "bottomright", sroc.legend.pos = "bottomleft", ...)
}
\arguments{
\item{fit.reitsma}{An object of class \code{reitsma} returned by fitting the
Reitsma bivariate model using the function named \code{reitsma} in the mada
package}

\item{show.data}{A logical indicating whether or not to show the mean
sensitivity and specificity of the individual studies}

\item{show.prediction}{A logical indicating whether or not draw a prediction region}

\item{summary.legend.pos}{A character string indicating the position of the
legend for the summary point and data points. This is passed to the \code{legend} function
and must be one of 'bottomleft', 'bottomright' (default), 'topleft',
'topright'.}

\item{sroc.legend.pos}{A character string indicating the position of the
legend for the SROC curve. This is passed to the \code{legend} function
and must be one of 'bottomleft' (default), 'bottomright', 'topleft',
'topright'.}

\item{...}{Additional arguments passed to the underlying plot method}
}
\description{
Draw a Summary ROC curve based on the results of the bivariate model
}
\details{
This function will draw a summary ROC curve based on a fit of the
  Reitsma bivariate model. The plot will always include a summary point with
  a confidence ellipse and a SROC curve. By default the plot will include the
  data points from the individual studies and a 95% prediction region around
  the summary point. These can be excluded by setting the arguments show.data
  or show.prediction to \code{FALSE}

  The argument extrapolate may also be set to TRUE or FALSE to indicate
  whether to extrapolate the SROC curve beyond the observed data points. The
  default is FALSE indicating no extrapolation
}
\seealso{
\code{\link[mada]{reitsma-class}}
}
