% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{reportDACC}
\alias{reportDACC}
\title{Summarise diagnostic accuracy data in a convenient format}
\usage{
reportDACC(df)
}
\arguments{
\item{df}{A data frame of diagnostic accuracy data as returned by
\code{extractDACC}}
}
\value{
A data frame summarising the diagnostic accuracy information for each
  study
}
\description{
Summarise diagnostic accuracy data in a convenient format
}
\seealso{
\code{extractDACC}, \code{\link[mada]{madad}}
}

