% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{runDAcc}
\alias{runDAcc}
\title{Perform meta-analysis of diagnostic accuracy data using the Reitsma bivariate
model}
\usage{
runDAcc(df, file, test_id, analysis_case = "Default", ...)
}
\arguments{
\item{df}{A data frame with a column of study IDs and integer variables for
observed frequencies of true positives, false negatives, false positives
and true negatives. The names of the columns should be StudyID, TP, FN, FP
and TN.}

\item{file}{A character string specifying the base directory for the
analysis. This function will create a subdirectory 'Results' within this
directory will be a 'Figures' subdirectory.}

\item{test_id}{A character string indicating the identity of the diagnostic
test}

\item{analysis_case}{A character string indicating a short name that
identifies this analysis. This is used in filenames so keep it short.}
}
\description{
Perform meta-analysis of diagnostic accuracy data using the Reitsma bivariate
model
}
