% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct.R
\name{runDirect}
\alias{runDirect}
\title{Run direct head to head meta-analysis for all possible pairwise comparisons
in a dataset}
\usage{
runDirect(df, treatments = NULL, file, data_type, effect_code, outcome,
  effect_measure, back_calc = FALSE, forest_plot = TRUE,
  show_fixed = TRUE, show_random = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} This should be in one of two formats. Arm level
data must contain the columns 'study' and 'treatment' where study is a
study id number (1, 2, 3 ...) and treatment is a treatment id number. If
the data are binary then the data frame should also contain columns
'responders' and 'sampleSize' reporting the total number of events and
total number analysed respectively for each arm. Relative effect data (e.g.
log odds ratio, log rate ratio) must contain the same study and treatment
columns plus the columns 'diff' and 'std.err'. Set diff=NA for the baseline
arm. The column std.err should be the standard error of the relative effect
estimate. For trials with more than two arms set std.err as the standard
error of the baseline arm. This determines the covariance which is used to
adjust for the correlation in multiarm studies.}

\item{treatments}{An optional (but recommended) data frame describing the
treatments. If present then this data frame should contain the columns
'id', 'description', 'shortname'. id gives the ID number of each treatment,
description gives the full name of each treatment, shortname gives a
shortened version of the name for each treatment. shortname is used for
labels and filenames in forest plots which is the main purpose of this
argument. If this is not provided then forest plots will be named
'Comparison1', Comparison2 etc.}

\item{file}{A character string specifying the directory where the results
will be saved. This function will create two new subdirectories
'Results/Direct' and 'Results/Direct/Figures' to store the output. These
are only created if they do not already exist}

\item{data_type}{A character string specifying which type of data has been
provided. Currently only 'treatment difference' or 'binary' are supported}

\item{effect_code}{A character string indicating the underlying effect
estimate. This is used to set the \code{sm} argument of the underlying
analysis functions from the \code{meta} package. Acceptable values are
'RD', 'RR', 'OR', 'HR', 'ASD', 'MD', 'SMD'}

\item{outcome}{A character string indicating what outcome we are currently
looking at. This is mainly used to automatically generate meaningful file
names for output so keep it short.}

\item{effect_measure}{A character string indicating what type of effect
measure is used, e.g. 'Rate Ratio', 'Odds Ratio' etc. This is used as a
label in forest plots so keep it short.}

\item{back_calc}{A logical indicating whether results should be back transformed.
This is used to set the corresponding \code{backtransf} argument of the
underlying functions from the \code{meta} package. If
\code{backtransf=TRUE} then log odds ratios (or hazard ratios etc) will
be converted to odds ratios on plots and print outs. Default is FALSE}

\item{forest_plot}{A logical indicating whether a forest plot should be
produced for each pairwise comparison.}

\item{show_fixed, show_random}{Logical indicating whether fixed effect
and/or random effects results should be shown on the forest plot. By
default both are TRUE and both results are shown. Set the appropriate
argument to FALSE if you want to exclude that result from the plot.}
}
\value{
A data frame containing the results of the meta-analysis
}
\description{
Run direct head to head meta-analysis for all possible pairwise comparisons
in a dataset
}
\details{
This function runs direct head to head meta-analysis for all
  possible pairwise comparisons in a dataset. The objective is to provide a
  minimal set of options and abstract as much of the technical detail as
  reasonably possible. This function calls a series of internal functions
  to do the work (linked below). These functions may accessed directly if
  additional flexibility is required but the intention is that this wrapper
  should cover the majority of common applications. The actual
  meta-analysis depends on functions from the \code{meta} package as does
  drawing forest plots. Currently only treatment differences (OR, HR, RR
  etc) and binary data are supported. Additional data types will be added
  in the future.

  The basic order of events is:
  \itemize{
   \item Run all possible pairwise comparisons in the dataset.
   \item Draw forest plots if requested (see \code{forest_plot} above).
   These are saved as jpg files in a subfolder named figures.
   \item Extract the key information from the meta-analysis output and save
   this as an excel file
  }
}
\seealso{
\code{\link{formatDataToDirectMA}}, \code{\link{doDirectMeta}},
  \code{\link{drawForest}}, \code{\link{extractDirectRes}}
}

