% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{runMTC}
\alias{runMTC}
\title{Run a mixed treatment comparison (MTC, a.k.a network meta-analysis)}
\usage{
runMTC(df, file, data_type, treatmentID, effect_measure, toi,
  analysis_set = "Default", outcome, analysis_case, doFixed = TRUE,
  doRandom = TRUE, model_type = "consistency", max_val = 5,
  prior = mtc.hy.prior("std.dev", "dunif", 0, max_val), burn_in = 10000,
  iterations = 20000, save_convergence = TRUE, back_calc = FALSE,
  includes_placebo = FALSE, placebo_code, report_order = "default",
  check_inconsistency = TRUE, rank = FALSE, higher_better = FALSE,
  keep_coda = FALSE, summarise_coda = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} This should be in one of two formats. Arm level
data must contain the columns 'study' and 'treatment' where study is a
study id number (1, 2, 3 ...) and treatment is a treatment id number. If
the data are binary then the data frame should also contain columns
'responders' and 'sampleSize' reporting the total number of events and
total number analysed respectively for each arm. Relative effect data (e.g.
log odds ratio, log rate ratio) must contain the same study and treatment
columns plus the columns 'diff' and 'std.err'. Set diff=NA for the baseline
arm. The column std.err should be the standard error of the relative effect
estimate. For trials with more than two arms set std.err as the standard
error of the baseline arm. This determines the covariance which is used to
adjust for the correlation in multiarm studies.}

\item{file}{A character string specifying the directory where the results
will be saved. This function will create a subdirectory 'Results/MTC'
within this directory will be one or more further directories
'Results/MTC/FixedEffects' and 'Results/MTC/RandomEffects' depending on
the values of \code{doFixed} and \code{doRandom}. These are only created
if they are required and do not exist already. Each results directory
will also contain a 'Figures' subdirectory.}

\item{data_type}{A character string specifying which type of data has been
provided. Must be one of 'treatment difference', 'binary' or 'continuous'.}

\item{treatmentID}{A data frame with columns 'description' defining the
treatment names and 'id' defining the treatment ID numbers. This is used
to set the \code{treatments} argument of the \code{mtc.network} function
from \code{gemtc}. An optional third column 'Order' may also be provided.
If present this should be a column of integers indicating the order in
which the interventions should be presented in output tables and figures}

\item{effect_measure}{A character string indicating what type of effect
measure is used, e.g. 'Rate Ratio', 'Odds Ratio', 'Risk Ratio' etc. This is
also used as a label in forest plots so keep it short.}

\item{toi}{A named vector specifying the treatments of interest. The names
should be the treatment names. The values should be the ID number of the
treatments in the network.}

\item{analysis_set}{A character string indicating which analysis set this
is. This is useful if there are multiple different sets of comparators in
a single project. If nothing is provided this will be set to the string
'Default'. This is used to set the \code{description} argument of the
underlying \code{mtc.network} function from the \code{gemtc} package}

\item{outcome}{A character string indicating which outcome is being analysed.}

\item{analysis_case}{A character string indicating what the current analysis
case is e.g. 'Base Case'}

\item{doFixed, doRandom}{Logical indicating which model is required. Fixed
effect, random effect or both. The default is to run both. Set the
appropriate argument to FALSE if you do not want to run that analysis.}

\item{model_type}{A character string describing the type of model. This is
passed to the \code{type} argument of the underlying \code{mtc.model}
function from \code{gemtc}. The underlying function accepts the values
"consistency", "nodesplit", "ume", or "use". Only consistency has been
tested in the context of this function.}

\item{max_val}{Used to set \code{om.scale} and create vague priors for the
between trials heterogeneity. For the log odds-ratio, values between 2
and 5 are considered reasonable. Default = 5.}

\item{prior}{A function call to \code{mtc.hy.prior} which specifies the type and
parameters of the prior distribution for the between trials
heterogeneity. This is passed directly to the \code{hy.prior} argument of
the underlying \code{mtc.model} function from the \code{gemtc} package.
The default is a uniform distribution on the between trials standard
deviation with a range from 0-5. This is adequate for log odds ratios,
log hazard ratios etc but should be changed if you are working with
continuous outcomes such as mean difference. For more details see
\code{\link[gemtc]{mtc.model}} and \code{\link[gemtc]{mtc.hy.prior}}}

\item{burn_in}{The number of iterations for the MCMC burn in period
(Default=10000)}

\item{iterations}{The number of iterations from the MCMC simulation to be
retained to estimate the results (Default=20000)}

\item{save_convergence}{Logical indicating whether run and keep plots to
assess the convergence of the MCMC chains. If TRUE (Default) then this
plots will be produced for the trace of each chain, autocorrelation
within each chain and the posterior distribution of each model parameter.}

\item{back_calc}{A logical indicating whether results should be back transformed.
If set to TRUE then log odds ratios (or hazard ratios etc) will
be converted to odds ratios on plots and print outs. Default is FALSE}

\item{includes_placebo}{Logical indicating whether the network includes
placebo. This requires slightly different handling internally as placebo
is always the comparator and never the intervention. Default is FALSE. If
this is set to TRUE then it is essential to also set the
\code{placebo_code} argument appropriately.}

\item{placebo_code}{An integer specifying the ID number of the placebo
treatment in the network if a placebo is included. Often placebo is given
the number 1 but not always.}

\item{report_order}{A character string indicating whether the treatments
should be reported in the order the order they were provided (default) or
if a custom order is required. Acceptable values are 'default' or
'custom'. If this is set to 'custom' then \code{treatmentID} must contain a
column named 'Order', see above.}

\item{check_inconsistency}{A logical indicating whether checks for
inconsistency should be performed. If set to \code{TRUE} (default) then the
\code{mtc.nodesplit} function from gemtc is used to run all possible
nodesplitting models for the current network.}

\item{rank}{A logical indicating whether ranking probabilities should be calculated.}

\item{higher_better}{A logical indicating whether higher or lower values
equate to a better outcome. If set to \code{TRUE} then higher values are
considered a better outcome. This is only used in the calculation of ranking
probabilities via \code{\link[gemtc]{rank.probability}}}

\item{keep_coda}{A logical indicating whether the coda should be extracted from
the MCMC simulation. This is set to \code{FALSE} by default. If set to
\code{TRUE} then the coda will be extracted and saved in folder named Coda}

\item{summarise_coda}{A logical indicating whether the coda should be
averaged across chains for multiple chains. This is set to \code{FALSE} by
default and results are returned for each chain separately. If set to
\code{TRUE} then the result will be the mean across all chains for each
variable at each iteration of the chains}
}
\description{
Run a mixed treatment comparison (MTC, a.k.a network meta-analysis)
}
\details{
This function provides an interface to run MTC analyses using the
  gemtc package. Although the function has a large number of arguments the
  default values should produce reasonable results in the majority of cases.
  The aim is to abstract away as much of the technical detail as possible so
  that reasonable values are chosen based on the type of data provided.

  The output of the analysis is automatically rearranged and saved in a
  'report ready' form to save time in downstream processing of the results
}
\seealso{
\code{\link[gemtc]{mtc.network}}, \code{\link[gemtc]{mtc.model}},
  \code{\link[gemtc]{mtc.run}}, \code{\link[gemtc]{mtc.run}},
  \code{\link{extractMTCResults}}, \code{\link{plotEstimates}}
}

