% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{saveDiagnostics}
\alias{saveDiagnostics}
\title{Save convergence diagnostics}
\usage{
saveDiagnostics(mtc, directory = "./ConvergenceDiagnostics")
}
\arguments{
\item{mtc}{An object of class \code{mtc.result} as returned by \code{mtc.run}
in the \code{gemtc} package}

\item{directory}{A character string indicating the directory to save the
results. By default a directory named ConvergenceDiagnostics will be
created as a subfolder of the current directory}
}
\description{
Save convergence diagnostics
}
\details{
Create pdf files of some basic diagnostic plots to check the
  convergence of MCMC models. Plots are produced by the ggmcmc package for
  the trace of the MCMC chain(s), the posterior density of the model
  parameters and the autocorrelation within chains
}
\seealso{
\code{\link[ggmcmc]{ggmcmc}}
}

