% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{saveModelCode}
\alias{saveModelCode}
\title{Extract JAGS model code from a \code{mtc.results} object}
\usage{
saveModelCode(mtcRes, modelFile)
}
\arguments{
\item{mtcRes}{An object of class \code{mtc.result} as returned by mtc.run
from the gemtc package}

\item{modelFile}{A character string specify the file path where the code will
be saved.}
}
\description{
Extract JAGS model code from a \code{mtc.results} object
}
\details{
The \code{mtc.run} function automatically generates JAGS code to
 specify a network meta-analysis model. This function extracts the JAGS code
 from the \code{mtc.results} object and saves it to a file to have a record of what
 was done.
}
\seealso{
\code{\link[gemtc]{mtc.run}}
}

